Melee Implementation
Just have to mess with values to get positioning and rotation right

+++ Assets/Scripts/Items/WeaponClass.cs
@@ -9,5 +9,5 @@ public class WeaponClass : ItemClass {
     public LayerMask whatToHit;
 
     public int damage = 0;
-    public int attackSpeed = 0;
+    public float attackSpeed = 0;
 }
